Mapcontrol loader section 1 written: L is NpcValue, Npc's ctor index is int
The loader's unidentified local is `NpcValue` (src/Npcvalue.h, 0x7c bytes): its
members at +4/+0x3c/+0x5c and its destructor's order (talk vector, drop vector,
name) match FUN_004a921c instruction-for-instruction, and its field offsets
account for every `L.fNN` the disassembly showed (+0xc boss, +0xe child, +0x10
npc_type, +0x1a min_damage, +0x1c max_damage, +0x1e accuracy, +0x20 evade, +0x22
armor, +0x2a element_weakness, +0x2c element_weakness_damage). The fill call is
`NpcValue npc_value = NpcValues::GetNpc(npc_values, id);` -- FUN_004a8858 is
NpcValues::GetNpc, already byte-exact and declared at src/Npcvalues.h:110; the
by-value return is why its argument list looks like (L, table, id).

Every other unnamed helper resolves to an already-reconstructed function:
NpcPtrVector_Count (0x4810a0), Npc::Npc (0x47a868), Map_AddNpc (0x4844c0),
Map_NpcIter_End (0x45ddfc), Pub_DecodeNumber_Map (0x486c38). The section delete
at 0x55949c is an RTL AnsiString helper.

Section 1 (the NPC section) is now written: npc_count decode, the inner loop,
`NpcPtrVector_Count+1` as the index, the 7-argument Npc constructor, the
NpcValue field copies, the 7-entry element-weakness table, and Map_AddNpc. The
alignment prefix now extends past reference index 775 (was 580).

Verified signature fix this unblocked: `Npc`'s constructor takes its index as
`int`, not `short` -- the reference's call site stores the argument with a 4-byte
`mov dword ptr [slot],eax`, where a short truncates to a 2-byte store. Changed in
src/Npc.h and src/Npc.cpp together (the only caller is the loader); the ctor body
stays byte-exact at 115/115 and the mangled name becomes @@Npc@$bctr$qissssis.

Npccontrol: NpcControl_Tick's frame now matches exactly (-236): the surplus slot
was a second `Player **player` iterator -- the reference reuses one across the
broadcaster and respawn loops. With the branch order, scan-loop assignment order
and party-loop guard order also pinned, 912 -> 788 mismatched and 51 -> 31 hunks.

Packets: Refresh_BuildReply transcribed (1394 ref / 1372 our) -- a deep cascade
from one 4-byte local (ref -256 vs our -252) that is not sit_state, cheat_x/y or
count; Paperdoll_BuildReply remains one try-wrap boundary away.

Sheet: byte-exact steady at 1602/1809; mismatched 11 -> 12.

+++ src/Npc.cpp
@@ -5,7 +5,7 @@
 
 #pragma package(smart_init)
 
-Npc::Npc(short npc_index,
+Npc::Npc(int npc_index,
          short npc_id,
          short x,
          short y,
